[TEAM2-2] Add : NCP SENS 문자 메시징 기능 추가
