--- src/main/java/com/rubycon/rubyconteam2/core/config/jpa/JpaConfig.java
@@ -1,9 +1,0 @@
-package com.rubycon.rubyconteam2.core.config.jpa;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfig {
-}
