+++ src/main/java/com/rubycon/rubyconteam2/global/config/jpa/JpaConfig.java
@@ -1,0 +1,9 @@
+package com.rubycon.rubyconteam2.global.config.jpa;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaConfig {
+}
