+++ src/main/java/com/rubycon/rubyconteam2/global/config/oauth/usertypes/FacebookOAuth2User.java
@@ -1,0 +1,59 @@
+package com.rubycon.rubyconteam2.global.config.oauth.usertypes;
+
+import com.rubycon.rubyconteam2.domain.user.domain.User;
+import lombok.Getter;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+import org.springframework.security.oauth2.core.user.OAuth2UserAuthority;
+
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
+
+@Getter
+public class FacebookOAuth2User implements OAuth2User {
+
+    private String id;
+    private String name;
+    private FacebookFicture picture;
+
+    public OAuth2User build(User user) {
+        this.id = user.getOauthKey();
+        this.name = user.getName();
+        this.picture = new FacebookOAuth2User.FacebookFicture().build(user);
+
+        return this;
+    }
+
+    @Override
+    public Map<String, Object> getAttributes() {
+        Map<String, Object> attrs = new HashMap<>();
+        attrs.put("id", this.id);
+        attrs.put("name", this.name);
+        attrs.put("image", this.picture.data.get("url"));
+
+        return attrs;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return Collections.singleton(new OAuth2UserAuthority(getAttributes()));
+    }
+
+    @Override
+    public String getName() {
+        return this.id;
+    }
+
+    @Getter
+    public static class FacebookFicture {
+        private Map<String, Object> data;
+
+        public FacebookFicture build(User user) {
+            data = new HashMap<>();
+            data.put("url", user.getProfileImage());
+            return this;
+        }
+    }
+}
