+++ src/main/java/com/rubycon/rubyconteam2/infra/sms/NCPMessage.java
@@ -1,0 +1,11 @@
+package com.rubycon.rubyconteam2.infra.sms;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class NCPMessage {
+    private String to;
+    private String contents;
+}
