+++ src/main/resources/application-config.yml
@@ -1,0 +1,32 @@
+spring:
+  security:
+    oauth2:
+      client:
+        registration:
+          google:
+            client-id: ${GOOGLE_CLIENT_ID:GOOGLE_CLIENT_ID_STRING}
+            client-secret: ${GOOGLE_CLIENT_SECRET:GOOGLE_CLIENT_SECRET_STRING}
+            scope: profile,email
+          facebook:
+            client-id: ${FACEBOOK_CLIENT_ID:FACEBOOK_CLIENT_ID_STRING}
+            client-secret: ${FACEBOOK_CLIENT_SECRET:FACEBOOK_CLIENT_SECRET_STRING}
+            scope: public_profile,email
+        provider:
+          facebook:
+            userInfoUri: https://graph.facebook.com/me?fields=id,name,email,picture
+
+custom:
+  oauth2:
+    kakao:
+      client-id: ${KAKAO_CLIENT_ID:KAKAO_CLIENT_ID_STRING}
+      client-secret: ${KAKAO_CLIENT_SECRET:KAKAO_CLIENT_SECRET_STRING}
+
+app:
+  jwt:
+    token-secret: ${JWT_SECRET:JWT_SECRET_STRING}
+    token-expiration-Msec: 864000000
+  sms:
+    ncp:
+      access-key: ${NCP_ACCESS_KEY:NCP_ACCESS_KEY_STRING}
+      secret-key: ${NCP_SECRET_KEY:NCP_SECRET_KEY_STRING}
+      service-id: ${NCP_SERVICE_ID:NCP_SERVICE_ID_STRING}
