--- src/main/resources/application-oauth.yml
@@ -1,27 +1,0 @@
-spring:
-  security:
-    oauth2:
-      client:
-        registration:
-          google:
-            client-id: ${GOOGLE_CLIENT_ID:GOOGLE_CLIENT_ID_STRING}
-            client-secret: ${GOOGLE_CLIENT_SECRET:GOOGLE_CLIENT_SECRET_STRING}
-            scope: profile,email
-          facebook:
-            client-id: ${FACEBOOK_CLIENT_ID:FACEBOOK_CLIENT_ID_STRING}
-            client-secret: ${FACEBOOK_CLIENT_SECRET:FACEBOOK_CLIENT_SECRET_STRING}
-            scope: public_profile,email
-        provider:
-          facebook:
-            userInfoUri: https://graph.facebook.com/me?fields=id,name,email,picture
-
-custom:
-  oauth2:
-    kakao:
-      client-id: ${KAKAO_CLIENT_ID:KAKAO_CLIENT_ID_STRING}
-      client-secret: ${KAKAO_CLIENT_SECRET:KAKAO_CLIENT_SECRET_STRING}
-
-app:
-  jwt:
-    tokenSecret: ${JWT_SECRET:JWT_SECRET_STRING}
-    tokenExpirationMsec: 864000000
